requirements.txt
aws-cdk-lib==2.29.1
constructs>=10.0.0,<11.0.0
python-dotenv == 0.20.0
